fix(templates): move component category grid into Design Documentation
The component overview page (hero + category grid) is the Design
Documentation template. Overview is now just the hero landing page.

Made-with: Cursor
