+++ packages/cli/templates/pages/documentation-design/page.tsx
@@ -9,13 +9,11 @@ import {
   XDSSideNavItem,
   XDSSideNavSection,
 } from '@xds/core/SideNav';
-import {XDSHeading, XDSText} from '@xds/core/Text';
+import {XDSText} from '@xds/core/Text';
+import {XDSButton} from '@xds/core/Button';
 import {XDSCard} from '@xds/core/Card';
-import {XDSCodeBlock} from '@xds/core/CodeBlock';
 import {XDSHStack, XDSVStack, XDSStackItem} from '@xds/core/Stack';
 import {XDSLayout, XDSLayoutContent} from '@xds/core/Layout';
-import {XDSDivider} from '@xds/core/Divider';
-import {XDSTable, pixel} from '@xds/core/Table';
 import {XDSGrid} from '@xds/core/Grid';
 import {radiusVars} from '@xds/core/theme/tokens.stylex';
 
@@ -30,12 +28,58 @@ const styles = stylex.create({
 // Data
 // ---------------------------------------------------------------------------
 
-const FOUNDATION_SECTIONS = [
-  {key: 'typography', name: 'Typography', desc: 'A type scale from display-1 down to supporting text, with weight and color options. Establishes visual hierarchy without guessing font sizes.'},
-  {key: 'colors', name: 'Colors', desc: 'A semantic palette that adapts across themes. Use surface, text, border, and accent tokens instead of raw hex values so your UI stays consistent.'},
-  {key: 'spacing', name: 'Spacing', desc: 'A 4px-based scale (0–10) used for padding, margins, and gaps. Keeps layouts aligned to a consistent rhythm across every component and page.'},
-  {key: 'shape', name: 'Shape', desc: 'Border radius tokens from sharp (2px) to fully rounded (pill). Controls, cards, and containers each have a designated radius so shapes feel intentional.'},
-  {key: 'icons', name: 'Icons', desc: 'A consistent icon set built for XDS components. Semantic and non-semantic color options with multiple sizes.'},
+const COMPONENT_CATEGORIES = [
+  {
+    label: 'Core',
+    items: [
+      {key: 'appshell', name: 'AppShell', desc: 'AppShell provides a foundational page layout with header, sidebar, and content regions. Use it to establish consistent structure across your application.'},
+      {key: 'avatar', name: 'Avatar', desc: 'Avatars represent a person or entity with an image, initials, or icon. They are commonly used in user profiles, comments, and contact lists.'},
+      {key: 'badge', name: 'Badge', desc: 'Badges display small counts or status labels. They can be attached to icons, buttons, or list items to surface key information at a glance.'},
+      {key: 'banner', name: 'Banner', desc: 'Banners show important, non-modal messages at the top of a page or section. They communicate status, warnings, or promotional information.'},
+      {key: 'button', name: 'Button', desc: 'Buttons let people take action. They can be used in forms, dialogs, and toolbars, or as standalone links.'},
+      {key: 'calendar', name: 'Calendar', desc: 'Calendar provides a date-picking grid for selecting single dates or date ranges. It integrates with form fields for date input.'},
+      {key: 'dialog', name: 'Dialog', desc: 'Dialogs are modal overlays that require user attention or action before continuing. They are used for confirmations, forms, and critical decisions.'},
+      {key: 'dropdownmenu', name: 'DropdownMenu', desc: 'DropdownMenu presents a list of actions or options in a floating overlay. It is triggered by a button and supports nested submenus.'},
+      {key: 'emptystate', name: 'EmptyState', desc: 'EmptyState provides a placeholder when there is no content to display. It guides users with a message, illustration, and optional call-to-action.'},
+      {key: 'hovercard', name: 'HoverCard', desc: 'HoverCard shows a rich preview of content when users hover over a trigger element. It is ideal for previewing profiles, links, or details.'},
+      {key: 'icon', name: 'Icon', desc: 'Icons are small visual symbols that represent actions, objects, or concepts. They improve scannability and reinforce meaning alongside text.'},
+      {key: 'kbd', name: 'Kbd', desc: 'Kbd renders keyboard shortcut hints in a styled inline element. Use it to show users which key combinations perform specific actions.'},
+      {key: 'link', name: 'Link', desc: 'Links provide navigation between pages or to external resources. They follow accessible anchor semantics with visual affordance.'},
+      {key: 'list', name: 'List', desc: 'List displays a vertical set of related items. It supports selection, icons, and metadata for building menus, nav lists, and more.'},
+      {key: 'popover', name: 'Popover', desc: 'Popover displays rich content in a floating panel anchored to a trigger element. It is used for forms, filters, and contextual tools.'},
+      {key: 'table', name: 'Table', desc: 'Table displays structured data in rows and columns with support for sorting, selection, and custom cell rendering.'},
+      {key: 'token', name: 'Token', desc: 'Tokens display compact metadata labels such as tags, categories, or filters. They can be dismissible and support selection state.'},
+      {key: 'tooltip', name: 'Tooltip', desc: 'Tooltips show concise helper text when users hover over or focus an element. They clarify icons, truncated labels, and controls.'},
+    ],
+  },
+  {
+    label: 'Layout',
+    items: [
+      {key: 'card', name: 'Card', desc: 'Cards group related content and actions in a contained surface. They can include headers, media, body text, and action bars.'},
+      {key: 'divider', name: 'Divider', desc: 'Dividers separate content into distinct sections with a subtle or strong horizontal line. They can optionally include a label.'},
+      {key: 'grid', name: 'Grid', desc: 'Grid provides a CSS grid-based layout container with configurable columns, rows, and gap. It simplifies responsive multi-column designs.'},
+      {key: 'stack', name: 'Stack', desc: 'Stack arranges child elements in a row or column with consistent gap spacing. It is the primary tool for one-dimensional layout composition.'},
+    ],
+  },
+  {
+    label: 'Navigation',
+    items: [
+      {key: 'breadcrumbs', name: 'Breadcrumbs', desc: "Breadcrumbs show the user's current location within a navigation hierarchy. They provide quick links back to parent pages."},
+      {key: 'sidenav', name: 'SideNav', desc: 'SideNav renders a vertical navigation panel with links, sections, and collapsible groups. It is used as the primary nav in dashboard layouts.'},
+      {key: 'tablist', name: 'TabList', desc: 'TabList switches between content views using a horizontal row of tabs. Only one tab is active at a time, and content changes without a page reload.'},
+      {key: 'topnav', name: 'TopNav', desc: 'TopNav provides an app-level navigation bar across the top of the page. It holds branding, primary links, search, and user actions.'},
+    ],
+  },
+  {
+    label: 'Form',
+    items: [
+      {key: 'checkboxinput', name: 'CheckboxInput', desc: 'CheckboxInput renders a single checkbox with a label. It is used for boolean opt-in choices like terms acceptance or feature toggles.'},
+      {key: 'selector', name: 'Selector', desc: 'Selector lets users pick a single item from a dropdown list. It supports search, grouping, and custom option rendering.'},
+      {key: 'switch', name: 'Switch', desc: 'Switch toggles a setting between on and off states with immediate effect. It is used for preferences, feature flags, and real-time controls.'},
+      {key: 'textinput', name: 'TextInput', desc: 'TextInput is a single-line text field for short user input like names, emails, and search queries. It supports icons, prefixes, and validation.'},
+      {key: 'typeahead', name: 'Typeahead', desc: 'Typeahead provides an autocomplete search input that suggests results as the user types. It supports async data sources and custom rendering.'},
+    ],
+  },
 ];
 
 // ---------------------------------------------------------------------------
@@ -43,79 +87,65 @@ const FOUNDATION_SECTIONS = [
 // ---------------------------------------------------------------------------
 
 function OverviewView({
-  onSelectSection,
+  onSelectComponent,
 }: {
-  onSelectSection: (key: string) => void;
+  onSelectComponent: (key: string) => void;
 }) {
   return (
     <XDSLayout contentWidth={1200} content={
       <XDSLayoutContent padding={8}>
         <XDSVStack gap={10}>
+          {/* Hero banner */}
           <XDSCard variant="cyan" padding={10}>
             <XDSHStack gap={8} vAlign="center">
               <XDSStackItem size="fill">
                 <XDSVStack gap={4}>
-                  <XDSText type="display-1">Design foundations</XDSText>
+                  <XDSText type="display-1">Web overview</XDSText>
                   <XDSText type="large" weight="normal" color="secondary">
-                    The design tokens and primitives that every component is
-                    built on. Use these foundations to create consistent,
-                    accessible interfaces.
+                    An open-source UI library to help developers quickly build
+                    beautiful, accessible products.
                   </XDSText>
+                  <XDSHStack>
+                    <XDSButton
+                      label="Get started"
+                      variant="primary"
+                      size="lg"
+                      onClick={() => onSelectComponent('getting-started')}
+                    />
+                  </XDSHStack>
                 </XDSVStack>
               </XDSStackItem>
               <XDSStackItem size="fill" />
             </XDSHStack>
           </XDSCard>
 
-          <XDSGrid columns={{minWidth: 260}} gap={8}>
-            {FOUNDATION_SECTIONS.map(item => (
-              <XDSVStack key={item.key} gap={3}>
-                <XDSCard
-                  variant="muted"
-                  padding={0}
-                  minHeight={160}
-                  xstyle={styles.previewCard}
-                  onClick={() => onSelectSection(item.key)}
-                />
-                <XDSVStack gap={0.5}>
-                  <XDSText type="body" weight="bold">
-                    {item.name}
-                  </XDSText>
-                  <XDSText type="body" color="secondary">
-                    {item.desc}
-                  </XDSText>
-                </XDSVStack>
-              </XDSVStack>
-            ))}
-          </XDSGrid>
-        </XDSVStack>
-      </XDSLayoutContent>
-    } />
-  );
-}
-
-function FoundationDetailView({activeNav}: {activeNav: string}) {
-  const item = FOUNDATION_SECTIONS.find(s => s.key === activeNav);
-  if (!item) return null;
-
-  return (
-    <XDSLayout contentWidth={960} content={
-      <XDSLayoutContent padding={8}>
-        <XDSVStack gap={8}>
-          <XDSVStack gap={2}>
-            <XDSText type="display-1">{item.name}</XDSText>
-            <XDSText type="body" color="secondary">{item.desc}</XDSText>
-          </XDSVStack>
-
-          <XDSDivider />
-
-          <XDSCard variant="muted" padding={10}>
-            <XDSVStack gap={3} hAlign="center" style={{textAlign: 'center'}}>
-              <XDSText type="body" color="secondary">
-                Documentation coming soon
-              </XDSText>
+          {/* Category sections */}
+          {COMPONENT_CATEGORIES.map(category => (
+            <XDSVStack key={category.label} gap={4}>
+              <XDSText type="display-2">{category.label}</XDSText>
+              <XDSGrid columns={{minWidth: 260}} gap={8}>
+                {category.items.map(item => (
+                  <XDSVStack key={item.key} gap={3}>
+                    <XDSCard
+                      variant="muted"
+                      padding={0}
+                      minHeight={160}
+                      xstyle={styles.previewCard}
+                      onClick={() => onSelectComponent(item.key)}
+                    />
+                    <XDSVStack gap={0.5}>
+                      <XDSText type="body" weight="bold">
+                        {item.name}
+                      </XDSText>
+                      <XDSText type="body" color="secondary">
+                        {item.desc}
+                      </XDSText>
+                    </XDSVStack>
+                  </XDSVStack>
+                ))}
+              </XDSGrid>
             </XDSVStack>
-          </XDSCard>
+          ))}
         </XDSVStack>
       </XDSLayoutContent>
     } />
@@ -127,8 +157,6 @@ function FoundationDetailView({activeNav}: {activeNav: string}) {
 // ---------------------------------------------------------------------------
 
 export default function DesignDocumentationPage() {
-  const [activePage, setActivePage] = useState<string>('home');
-
   return (
     <XDSAppShell
       variant="section"
@@ -141,27 +169,23 @@ export default function DesignDocumentationPage() {
           <XDSSideNavSection title="Navigation" isHeaderHidden>
             <XDSSideNavItem
               label="Home"
-              isSelected={activePage === 'home'}
-              onClick={() => setActivePage('home')}
+              isSelected
             />
           </XDSSideNavSection>
-          <XDSSideNavSection title="Foundations">
-            {FOUNDATION_SECTIONS.map(item => (
-              <XDSSideNavItem
-                key={item.key}
-                label={item.name}
-                isSelected={activePage === item.key}
-                onClick={() => setActivePage(item.key)}
-              />
-            ))}
-          </XDSSideNavSection>
+
+          {COMPONENT_CATEGORIES.map(category => (
+            <XDSSideNavSection key={category.label} title={category.label}>
+              {category.items.map(item => (
+                <XDSSideNavItem
+                  key={item.key}
+                  label={item.name}
+                />
+              ))}
+            </XDSSideNavSection>
+          ))}
         </XDSSideNav>
       }>
-      {activePage === 'home' ? (
-        <OverviewView onSelectSection={setActivePage} />
-      ) : (
-        <FoundationDetailView activeNav={activePage} />
-      )}
+      <OverviewView onSelectComponent={() => {}} />
     </XDSAppShell>
   );
 }
